+++ app/src/main/java/com/julia/android/worderly/ui/game/view/GameActivityView.java
@@ -1,5 +1,5 @@
 package com.julia.android.worderly.ui.game.view;
 
 public interface GameActivityView {
-    void setChatTabTitle(int number);
+    void setChatTabTitle(long number);
 }
